fix warnings about missing prototypes, printf specifiers that don't match long parameters, unused or possibly uninitialized variables, and so on
git-svn-id: http://svn.eol.ucar.edu/svn/rdss/trunk/zebra/source@8287 3f9b6c58-9310-455a-aec6-8b30269ab333

+++ src/ds/dc_IRGrid.c
@@ -167,7 +167,8 @@ DataChunk *dc;
  */
 	printf ("IRGRID class, %d platforms\n", nplat);
 	for (plat = 0; plat < nplat; plat++)
-		printf ("\t%2d: (%s) at %.4f %.4f %.2f\n", pinfo[plat].pi_Id,
+		printf ("\t%2ld: (%s) at %.4f %.4f %.2f\n", 
+			(long)pinfo[plat].pi_Id,
 			ds_PlatformName (pinfo[plat].pi_Id),
 			pinfo[plat].pi_Loc.l_lat, pinfo[plat].pi_Loc.l_lon,
 			pinfo[plat].pi_Loc.l_alt);
